feat: arruma schema dos filtros e simplifica código dos adapters
