+++ src/features/acf/diabetes/backend/controller.ts
@@ -1,5 +1,6 @@
 
 import type { DiabetesAcfItem } from "../common/model"
+import type { FilterParams } from "../common/schema"
 import { diabetesFilterToDb, diabetesPageDbToModel } from "./adapter"
 import { diabetesListCount, diabetesPage } from "./repository"
 import type { FilterItem } from "@/services/lista-nominal/ListaNominal"
@@ -8,7 +9,7 @@ export const diabetesData = async(
     municipalitySusID: string,
     teamIne: string,
     page: number,
-    filters: FilterItem
+    filters: FilterParams
 ): Promise<DiabetesAcfItem[]> => {
     const filtersDb = diabetesFilterToDb(filters)
     const data = await diabetesPage(municipalitySusID, teamIne, page, filtersDb)
